Pull out GifDownloader class.
// FREEBIE
